\name{ANYSXP}
\alias{ANYSXP}
\alias{EXPRSXP}
\alias{EXPTRSXP}
\alias{INTSXP}
\alias{LANGSXP}
\alias{LGLSXP}
\alias{RAWSXP}
\alias{REALSXP}
\alias{S4SXP}
\alias{STRSXP}
\alias{SYMSXP}
\alias{VECSXP}
\alias{CPLXSXP}
\alias{ENVSXP}
\alias{CHARSXP}

\title{Enumerated constants for R data types}
\description{}
\references{
  These come from Rinternals.h
}